handy_forms/requirements.txt
django
djangorestframework
django-cors-headers
mysqlclient==2.0.0